Add avgmax_attention to validation
