+++ training/utils.py
@@ -4,6 +4,7 @@ import random
 
 from dataset.clip_generator import CHARACTER_ENCODING
 from training import BASE_SYMBOLS
+from training.tacotron2_model.utils import get_mask_from_lengths
 
 
 CHECKPOINT_SIZE_MB = 333
@@ -144,3 +145,32 @@ def check_early_stopping(validation_losses):
         if difference < EARLY_STOPPING_MIN_DIFFERENCE:
             return True
     return False
+
+
+def calc_avgmax_attention(mel_lengths, text_lengths, alignment):
+    """
+    Calculate Average Max Attention for Tacotron2 Alignment.
+    Roughly represents how well the model is linking the text to the audio.
+    Low values during training typically result in unstable speech during inference.
+    
+    Parameters
+    ----------
+    mel_lengths : torch.Tensor
+        lengths of each mel in the batch
+    text_lengths : torch.Tensor
+        lengths of each text in the batch
+    alignment : torch.Tensor
+        alignments from model of shape [B, mel_length, text_length]
+    
+    Returns
+    -------
+    float
+        average max attention
+    """
+    mel_mask = get_mask_from_lengths(mel_lengths, device=alignment.device)
+    txt_mask = get_mask_from_lengths(text_lengths, device=alignment.device)
+    attention_mask = txt_mask.unsqueeze(1) & mel_mask.unsqueeze(2)# [B, mel_T, 1] * [B, 1, txt_T] -> [B, mel_T, txt_T]
+    
+    alignment = alignment.data.masked_fill(~attention_mask, 0.0)
+    avg_prob = alignment.data.amax(dim=2).sum(1).div(mel_lengths.to(alignment)).mean().item() # [B, mel_T, txt_T]
+    return avg_prob
